feature/add-binary-code-generator

+++ src/CodeGenerator/function_table.h
@@ -10,9 +10,9 @@ class FunctionTable {
         FunctionTable();
         ~FunctionTable();
 
-        void add(std::string name, int address);
+        void add(std::string name, long address);
         std::string getName(int index);
-        int getAddress(std::string name);
+        long getAddress(std::string name);
         void printTable(void);
     private:
         std::vector<std::pair<std::string, int>> names;
